requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
decorator==4.4.2
Django==4.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-password-validators==1.7.0
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
Markdown==3.4.1
moviepy==1.0.3
numpy==1.23.4
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
Pillow==9.3.0
proglog==0.1.10
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.12
waitress==2.1.2
whitenoise==6.2.0